Position control and halogen intensity for Auto
This document consists implementing a digital PID controller using Labview software to control a DC motor and Arduino to control the intensity of halogens. Initially given a brief description of the drivers, then proceed to find the equation of the plant, PID controller design, implementation and verification of the results.